refactor(cfg): graph.py becomes the cfg package
Sixth of the stack. What is left of graph.py is graph structure, so it moves to
static_analyzer/cfg and splits along its three concerns: call_graph.py holds
CallGraph, edge.py the edge types, location_key.py the physical-location
identity used to dedup LSP aliases.

Git sees call_graph.py as a rename of graph.py, so the diff shows what actually
moved rather than 300 lines deleted and 300 added.

The move also removed the last cfg -> clustering import: CallGraph carried a
MethodClusterPaths import left over from before the partition moved off it, and
nothing used it. With that gone the layering is genuinely one-way — cfg knows
nothing about clustering, clustering imports cfg — and both import orders now
work rather than only the lucky one.
