+++ agents/llm_renderers/call_graph.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 from collections.abc import Sequence
 
 from static_analyzer.constants import NodeType
-from static_analyzer.graph import CallGraph
+from static_analyzer.cfg import CallGraph
 from static_analyzer.node import Node
 
 logger = logging.getLogger(__name__)
