+++ static_analyzer/incremental_orchestrator.py
@@ -26,7 +26,7 @@ from static_analyzer.engine.lsp_constants import EdgeStrategy
 from static_analyzer.engine.result_converter import convert_to_codeboarding_format
 from static_analyzer.engine.source_inspector import SourceInspector
 from static_analyzer.engine.utils import definition_location, uri_to_path
-from static_analyzer.graph import CallGraph
+from static_analyzer.cfg import CallGraph
 from static_analyzer.internal_references import parent_qualified_name
 from static_analyzer.node import Node
