+++ tests/agents/test_validation.py
@@ -24,7 +24,7 @@ from agents.agent_responses import (
     ComponentFiles,
     FileClassification,
 )
-from static_analyzer.graph import CallGraph
+from static_analyzer.cfg import CallGraph
 from static_analyzer.clustering import ClusterResult
 from static_analyzer.node import Node
 from static_analyzer.constants import NodeType
