+++ tests/health/test_runner.py
@@ -5,7 +5,7 @@ from health.models import HealthCheckConfig, StandardCheckSummary
 from health.runner import run_health_checks
 from static_analyzer.analysis_result import StaticAnalysisResults
 from static_analyzer.constants import Language
-from static_analyzer.graph import CallGraph
+from static_analyzer.cfg import CallGraph
 from static_analyzer.node import Node
