+++ tests/static_analyzer/scripts/diagnose_relations.py
@@ -23,7 +23,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parents[3]))
 
 from static_analyzer import get_static_analysis
 from static_analyzer.cluster_relations import build_component_relations
-from static_analyzer.graph import CallGraph
+from static_analyzer.cfg import CallGraph
 from utils import get_artifact_dir
 
 logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
