+++ tests/static_analyzer/test_call_site_locations.py
@@ -1,7 +1,7 @@
 """Tests for preserving call-site locations on static-analysis edges."""
 
 from static_analyzer.constants import NodeType
-from static_analyzer.graph import CallGraph, Edge
+from static_analyzer.cfg import CallGraph, Edge
 from static_analyzer.node import Node
