+++ tests/static_analyzer/test_global_relations.py
@@ -53,7 +53,7 @@ from static_analyzer.cluster_relations import (
     build_global_relations,
 )
 from static_analyzer.constants import NodeType
-from static_analyzer.graph import CallGraph, Edge
+from static_analyzer.cfg import CallGraph, Edge
 from static_analyzer.node import Node
