+++ tests/static_analyzer/test_graph.py
@@ -6,7 +6,7 @@ import networkx as nx
 from static_analyzer.constants import NodeType
 from static_analyzer.leiden_utils import find_partition
 from static_analyzer.node import Node
-from static_analyzer.graph import CallGraph, Edge
+from static_analyzer.cfg import CallGraph, Edge
 from static_analyzer.clustering import ClusterResult, ClusteringService
