+++ tests/static_analyzer/test_reference_resolver.py
@@ -7,7 +7,7 @@ from agents.agent_responses import AnalysisInsights, Component, Relation, Relati
 from agents.file_index_models import FileMethodGroup, MethodEntry
 from static_analyzer.analysis_result import StaticAnalysisResults
 from static_analyzer.constants import NodeType
-from static_analyzer.graph import Edge
+from static_analyzer.cfg import Edge
 from static_analyzer.node import Node
 from static_analyzer.reference_resolver import StaticReferenceResolver
